Refactored performance evaluation code
